refactor: split prompt
