+++ src/agents/action/llm.test.ts
@@ -4,7 +4,7 @@ import { beforeAll, describe, expect, it } from 'vitest'
 import { initBot, useBot } from '../../composables/bot'
 import { botConfig, initEnv } from '../../composables/config'
 import { initLogger } from '../../utils/logger'
-import { genSystemBasicPrompt } from '../../utils/prompt'
+import { generateSystemBasicPrompt } from '../prompt/llm-agent.plugin'
 import { initAgent } from './llm'
 
 describe('openAI agent', { timeout: 0 }, () => {
@@ -22,7 +22,7 @@ describe('openAI agent', { timeout: 0 }, () => {
       bot.bot.once('spawn', async () => {
         const text = await agent.handle(
           messages(
-            system(genSystemBasicPrompt('airi')),
+            system(generateSystemBasicPrompt('airi')),
             user('Hello, who are you?'),
           ),
           async (c) => {
