+++ src/agents/chat/llm.ts
@@ -4,8 +4,8 @@ import type { ChatHistory } from './types'
 import { useLogg } from '@guiiai/logg'
 import { system, user } from 'neuri/openai'
 
-import { genChatAgentPrompt } from '../../utils/prompt'
 import { toRetriable } from '../../utils/reliability'
+import { genChatAgentPrompt } from '../prompt/chat'
 
 const logger = useLogg('chat-llm').useGlobalConfig()
